Rename project to EntirePaper
