+++ lib/src/main/java/com/dwolfnineteen/entirepaper/entities/Change.java
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2023 DWolf Nineteen & The JPD-API contributors
+Copyright (c) 2023 DWolf Nineteen & The EntirePaper contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -19,12 +19,27 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package com.dwolfnineteen.jpdapi.exceptions;
+package com.dwolfnineteen.entirepaper.entities;
 
 import org.jetbrains.annotations.NotNull;
 
-public abstract class APIException extends RuntimeException {
-    public APIException(@NotNull String message) {
-        super(message);
+public class Change {
+    private String commit;
+    private String summary;
+    private String message;
+
+    @NotNull
+    public String getCommit() {
+        return commit;
+    }
+
+    @NotNull
+    public String getSummary() {
+        return summary;
+    }
+
+    @NotNull
+    public String getMessage() {
+        return message;
     }
 }
