+++ lib/src/main/java/com/dwolfnineteen/entirepaper/entities/Download.java
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2023 DWolf Nineteen & The JPD-API contributors
+Copyright (c) 2023 DWolf Nineteen & The EntirePaper contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -19,17 +19,24 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package com.dwolfnineteen.jpdapi.entities;
+package com.dwolfnineteen.entirepaper.entities;
 
 import com.google.gson.annotations.SerializedName;
-import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.NotNull;
 
-public class Error {
-    @SerializedName("error")
-    private String errorMessage;
+public class Download {
+    private String name;
 
-    @Nullable
-    public String getErrorMessage() {
-        return errorMessage;
+    @SerializedName("sha256")
+    private String SHA256;
+
+    @NotNull
+    public String getName() {
+        return name;
+    }
+
+    @NotNull
+    public String getSHA256() {
+        return SHA256;
     }
 }
