+++ lib/src/main/java/com/dwolfnineteen/entirepaper/exceptions/ProjectNotFoundException.java
@@ -19,7 +19,12 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-rootProject.name = "EntirePaper"
+package com.dwolfnineteen.entirepaper.exceptions;
 
-include("lib")
-include("samples")
+import org.jetbrains.annotations.NotNull;
+
+public class ProjectNotFoundException extends APIException {
+    public ProjectNotFoundException(@NotNull String projectID) {
+        super(String.format("Project '%s' not found", projectID));
+    }
+}
