+++ lib/src/main/java/com/dwolfnineteen/entirepaper/exceptions/VersionNotFoundException.java
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2023 DWolf Nineteen & The JPD-API contributors
+Copyright (c) 2023 DWolf Nineteen & The EntirePaper contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -19,13 +19,12 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package com.dwolfnineteen.jpdapi.entities;
+package com.dwolfnineteen.entirepaper.exceptions;
 
-import com.google.gson.annotations.SerializedName;
+import org.jetbrains.annotations.NotNull;
 
-public enum Channel {
-    @SerializedName("default")
-    DEFAULT,
-    @SerializedName("experimental")
-    EXPERIMENTAL
+public class VersionNotFoundException extends APIException {
+    public VersionNotFoundException(@NotNull String version) {
+        super(String.format("Version '%s' not found", version));
+    }
 }
